requirements.txt
torch==2.1.2
transformers==4.40.1
accelerate==0.29.3
bitsandbytes==0.43.1
safetensors==0.4.2
peft==0.11.1
pillow==10.3.0
opencv-python==4.9.0.80
requests==2.31.0
open-clip-torch==2.24.0
sentence-transformers==2.7.0
flash-attn==2.5.6 
